+++ CHANGELOG.md
@@ -9,5 +9,11 @@
 * Fixed the "Show Focus" option for the Headline View not working.
 * Fixed the Focus and mouseover highlights not being shown in Headline View.
 * Fixed a Lua error when selecting the "Crescent" texture for the Target or Focus highlight.
-* Updated integrated libraries (Ace3 r1414-alpha).
 * Fixed the options for showing friendly minions, pets, guardians, totems, and NPCs having no effect (checkboxes could not be checked) [Comment #8597].
+* Disabled the options for showing friendly pets, guardians, and totems if showing friendly players is disabled, as Blizzard does not show these nameplates then.
+* Updated integrated libraries (Ace3 r1414-alpha, LibDualSpec-1.0 v1.35.0, LibSharedMedia-3.0 v12.1.0).
+* Fixed a bug where the absorb bar briefly showed at full width when a new nameplate was displayed.
+* Fixed automatic role detection not recognizing Druids in Bear Form as tanks on WoW Forever.
+* Added support for WoW Forever's character surnames - full names (first and last) are now shown on nameplates, with a new option to show first names only.
+* Fixed a Lua error on WoW Forever for Paladins (and some other classes), caused by changes to the Blizzard API used for role detection [Comment #8694].
+* Enabled target of target names again on WoW Midnight and WoW Forever [Comment #8697].